requirements.txt
streamlit
fredapi
yfinance
pandas
python-dotenv
requests
beautifulsoup4
lxml
xlrd==1.2.0